Move GraphSlice to tslib/react
